Fixing FxCopAnalyzer warnings with unit tests
