+++ tests/SourceCode.Clay.AspNetCore.Middleware.Correlation.Tests/Testing/TestWebApplicationStartup.cs
@@ -7,6 +7,8 @@ namespace SourceCode.Clay.AspNetCore.Middleware.Correlation.Tests
 {
     public class TestWebApplicationStartup
     {
+#pragma warning disable CA1822 // Mark members as static - Called by runtime
+#pragma warning disable CA1801 // Remove unused parameter - Part of interface
         public void ConfigureServices(IServiceCollection services)
         {
             //services.AddCorrelationId();
@@ -21,5 +23,7 @@ namespace SourceCode.Clay.AspNetCore.Middleware.Correlation.Tests
                 await context.Response.WriteAsync("Hello world");
             });
         }
+#pragma warning restore CA1801 // Remove unused parameter
+#pragma warning restore CA1822 // Mark members as static
     }
 }
